Finish merging main

+++ glyphs.go
@@ -8,7 +8,7 @@ package main
 
 var iconGlyph int64 = 61440
 
-var glyphs = map[string]int{
+var glyphs = map[string]int64{
 	"car":               0xe83c,
 	"bus":               0xe91e,
 	"robot":             61566,
